refactor: put `PartialJob` to the other file
