+++ app/models/entities/job.rb
@@ -1,18 +1,9 @@
 # frozen_string_literal: true
 
 require 'dry-struct'
-require_relative '../gateways/reed_api'
 
 module Skiller
   module Entity
-    # Library for job information
-    class PartialJob < Dry::Struct
-      attribute :id, Types::Strict::String
-      attribute :title, Types::Strict::String
-      attribute :description, Types::Strict::String
-      attribute :location, Types::Strict::String
-    end
-
     # Job information from Reed Details API
     class Job < PartialJob
       attribute :min_year_salary, Types::Optional::Float
